Set the stage for LimeSDR testing
We're having trouble bringing up the LMS7002m in the XTRX, so let's
double-check bringup with a vanilla LimeSDR, using vendored versions of
LimeSuite/libLMS7Support.so, SoapySDR.jl, etc...

This commit adds two new targets to `software/Makefile`: it can
download/install `LimeSuite` and also `SoapySDR.jl`.  In the `scripts`
directory, we also have a new julia script that sets up a
`JuliaLocalPreferences.toml` file to point JLLs to use the `.so` files
we built for things like `libSoapySDR.so` and `libLMS7support.so`.