Remove `fpga-source` submodule
We don't need to use submodules to track fairwaves stuff since it's not
a moving target, and it's nice to not have to download it when we
download all the other submodules.

We still have the `make -C software fairwaves` target to download all
the fairwaves stuff, if you need it for local development.