Automatically set CGEN to 8x samplerate (#102)
This automatically sets `CGEN` to a default value when setting the
samplerate of a Tx or Rx stream.  Note that although the SoapySDR API
makes it seem like the Tx and Rx samplerates are independent, they are
rather tightly coupled by the LMS7002M chip.  While there is some
flexibility in the chip, we are likely never going to make use of a full
duplex operation where the samplerates are different, so I have opted to
take the simple approach and simplify the codebase to instead hint to
the developer (where allowed by the SoapySDR API) that there is actually
only a single TSP rate.

With this change we get much more reliably smooth looking transmissions,
although as the samplerate increases we still get a bunch of noise and
other issues.  I suspect improper reference clock tuning may be an issue
here, and so further experimentation will have to wait for that to get
sorted.